+++ x-pack/solutions/security/packages/kbn-cloud-security-posture/public/src/components/fleet_extensions/hooks/use_load_cloud_setup.ts
@@ -8,7 +8,6 @@ import { useCallback } from 'react';
 import semverGte from 'semver/functions/gte';
 import { PackagePolicyValidationResults } from '@kbn/fleet-plugin/common/services';
 import { NewPackagePolicy, NewPackagePolicyInput, PackageInfo } from '@kbn/fleet-plugin/common';
-import { SECURITY_SOLUTION_ENABLE_CLOUD_CONNECTOR_SETTING } from '@kbn/management-settings-ids';
 import { SetupTechnology } from '@kbn/fleet-plugin/common/types';
 import { IUiSettingsClient } from '@kbn/core/public';
 import { CloudSetup } from '@kbn/cloud-plugin/public';
@@ -20,6 +19,7 @@ import {
 } from '../utils';
 import { useSetupTechnology } from './use_setup_technology';
 import { CloudProviders, CloudSetupConfig, UpdatePolicy } from '../types';
+import { SECURITY_SOLUTION_ENABLE_CLOUD_CONNECTOR_SETTING } from '../constants';
 
 const getSelectedInput = (options: NewPackagePolicyInput[], defaultProviderType: string) => {
   // Looks for the enabled deployment (aka input). By default, all inputs are disabled.
